+++ packages/falso/src/lib/sports.ts
@@ -1,9 +1,11 @@
 import { fake, FakeOptions, getRandomInRange, randElement } from './core/core';
 import { data } from './sports.json';
 
+export type Category = 'olympic' | 'winterOlympic' | 'outdoor';
+
 export interface SportCategories extends FakeOptions {
   // This categories can be extended in the future
-  category?: 'olympic' | 'outdoor' | 'winterOlympic';
+  category?: Category;
 }
 
 const categoriesCount = Object.keys(data)?.length;