+++ packages/falso/src/tests/between-date.spec.ts
@@ -1,9 +1,12 @@
 import { randBetweenDate } from '../lib/between-date';
 
+const now = new Date();
+const future = new Date('12/02/2040');
+
 describe('randBetweenDate', () => {
   it('should return date between from and to', () => {
-    const from = new Date();
-    const to = new Date('12/02/2022');
+    const from = now;
+    const to = future;
 
     const result = randBetweenDate({ from, to });
 
@@ -12,8 +15,8 @@ describe('randBetweenDate', () => {
   });
 
   it('should use strings for dates', () => {
-    const from = new Date().toISOString();
-    const to = new Date('12/02/2022').toISOString();
+    const from = now.toISOString();
+    const to = future.toISOString();
 
     const result = randBetweenDate({ from, to });
 
@@ -26,8 +29,8 @@ describe('randBetweenDate', () => {
     let to: Date;
 
     beforeEach(() => {
-      from = new Date('12/02/2022');
-      to = new Date();
+      from = future;
+      to = now;
     });
 
     it('should throw error', () => {