+++ packages/falso/src/tests/incremental-number.spec.ts
@@ -48,9 +48,21 @@ describe('incrementalNumber', () => {
     );
   });
 
+  it('should not allow invalid range when negative step', () => {
+    expect(() => incrementalNumber({ from: 1, to: 5, step: -1 })).toThrowError(
+      '`to` should be lower or equal to `from`, for example: {from: 5, to: 1, step: -1}'
+    );
+  });
+
   it('should not allow step 0', () => {
     expect(() => incrementalNumber({ from: 1, to: 100, step: 0 })).toThrow(
       '`step` should be a number different than 0, for example: {from: 1, step: 1}'
+    );
+  });
+
+  it('should not allow negative from', () => {
+    expect(() => incrementalNumber({ from: -1, to: 100, step: 1 })).toThrow(
+      '`from` should be a number greater than 0'
     );
   });
 });